pyproject.toml
[project]
name = "jordandm_online_portfolio"
version = "0.1.0"
description = "A web portfolio for Jordan Morgan"
requires-python = ">=3.10"
dependencies = [
    "mcp[cli]>=1.2.0",
    "uvicorn",
    "fastapi",
    "httpx-sse",
    "pydantic>=2.0.0",
    "rich",
    "typer",
    "requests",
    "pyyaml",
    "django>=5.2.3",
    "pytest",
    "httpx",
    "pytest-asyncio",
    "whitenoise>=6.6.0",
    "openai",
    "djangorestframework",
    "python-dotenv",
    "gunicorn (>=23.0.0,<24.0.0)",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
asyncio_mode = "auto"

[project.optional-dependencies]
test= [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.0.0",
]
